Remove test_sshd_test_model
